src/extract/requirements.txt
requests==2.31.0
# beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.1.1